Move HistoricalCasesChart into folder and extract tooltip and scroller components
